+++ ScrapyKeeper/uwsgi.py
@@ -4,8 +4,8 @@
 # @Author    : modm
 '''
 you can start the server by uwsgi
-like gunicorn -w 4 SpiderKeeper.uwsgi:app
+like gunicorn -w 4 ScrapyKeeper.uwsgi:app
 '''
-from SpiderKeeper.app import app, initialize
+from ScrapyKeeper.app import app, initialize
 
 initialize()